requirements.txt
gradio
google-generativeai
scikit-learn
pandas
matplotlib
seaborn
joblib
python-dotenv
